sistema_impresion_bot/requirements.txt
# Telegram Bot
python-telegram-bot==20.7
python-dotenv==1.0.0

# Panel de Control (GUI) - Sin Pillow por ahora
customtkinter==5.2.0

# Base de Datos y Utilidades
pandas==2.2.0
python-dateutil==2.8.2

# Sonidos
pygame==2.5.2

# Reportes
reportlab==4.1.0
xlsxwriter==3.1.9

# Manejo de Archivos
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1